requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Jinja2==3.1.6
Werkzeug==3.1.3
WTForms==3.2.1
itsdangerous==2.2.0
MarkupSafe==3.0.2

SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
greenlet==3.2.1

Telethon==1.42.0
selenium==4.40.0
webdriver-manager==4.0.2
cloudinary==1.44.0
requests==2.32.5
urllib3==2.6.3
beautifulsoup4==4.14.3
google_search_results==2.4.2

customtkinter==5.2.2
darkdetect==0.8.0
python-dotenv==1.2.1
colorama==0.4.6
rich==14.2.0
tqdm==4.67.1
pillow==12.1.0
pandas==2.3.3
numpy==2.2.6

bcrypt==5.0.0
email-validator==2.3.0
validate_docbr==1.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11

